refactor:navbar login buttons updated
